2^24-1 would not fit in a uint 16

+++ server.go
@@ -233,8 +233,7 @@ func deserializeCertificate(c []byte) (*certificateBody, error) {
 func serializeCertificate(c *certificateBody) ([]byte, error) {
 	certListBody := make([]byte, 0, 0xffffff) // 2^24-1 is maximim length
 	for _, ci := range c.certificateList {
-		//XXX Actually, ALL certificates cant be bigger than this size
-		if uint16(len(ci)) > uint16(math.Pow(2, 24)-1) {
+		if uint32(len(ci)) > uint32(math.Pow(2, 24) - 1) {
 			return nil, errors.New("A certificate in the list has exceeded the size limiet of 2^24-1")
 		}
 		certificateLen := writeBytesFromUint24(uint32(len(ci)))
